+++ packages/internal_wiki/src/index.ts
@@ -1,3 +1,33 @@
 export { getTrafficSignsWiki, trafficSignsWiki, type WikiSign } from './getTrafficSignsWiki.js'
 export { getWikiSnapshotMeta, type WikiSnapshotMeta } from './getWikiSnapshotMeta.js'
-export { countryWikiConfigs, wikiSnapshotCountryPrefixes } from './countryWikiConfigs.js'
+export {
+  catalogueWikiConfigs,
+  catalogueWikiCountryPrefixes,
+  type CatalogueCountryConfig,
+  type CataloguePageConfig,
+} from './catalogueWikiConfigs.js'
+export {
+  countryWikiConfigs,
+  wikiSnapshotCountryPrefixes,
+  WIKI_BASE,
+  type CountryWikiConfig,
+  type WikiPageConfig,
+} from './countryWikiConfigs.js'
+export { fetchWikiPage } from './fetchWikiPage.js'
+export {
+  cleanWikiSignName,
+  dedupeWikiSigns,
+  extractDeTrafficSignValue,
+  extractTrafficSignId,
+  finalizeWikiSignName,
+  isWikiTaggingCell,
+  looksLikeWikiSignNameCell,
+  parseBelgiumTable,
+  parseDeRowIdTable,
+  parseUniversalTable,
+  parseWikiTags,
+  toWikiSign,
+  type ParsedDeWikiRow,
+  type ParsedWikiRow,
+} from './parseWiki/parseWikiTables.js'
+export { parseWikiHtml, generateWikiSnapshotForCountry } from './generateWikiSnapshots.js'